[docker]: move String() to docker.go for all platforms.
